Adding GUI when pressing TAB key
::Functionality::
When "Back" input is activated(TAB key for keyboard, Back button for
joypad), the "HELP" GUI pops up.
In the "HELP" GUI,  a figure for controller mapping and the task
objectives with time stamps are shown.

::Usage::
In the folder, Asset>Score Board>Prefab, instantiate the prefab "Help
GUI.prefab" to any place in the scene hierarchy.
In order for "Back" input to work properly, add the "Back" input axis in
the Edit>Project Settings>Input.

::APIs supported::
public void addObjective(string s)
Inside the prefab, the gameobject "EventSystem" has a script attached to
it, called "Score Board Manager".  Whenever this function is called, the
argument string describing the new objective is added to the history of
objectives and the GUI is updated.